WIP new deep merge tests
