+++ packages/records/lib/models/records.ts
@@ -10,8 +10,7 @@ import type {
     ReturnedRecord,
     UpsertSummary,
     MergingStrategy,
-    CursorOffset,
-    UnencryptedRecordData
+    CursorOffset
 } from '../types.js';
 import { decryptRecordData, encryptRecords } from '../utils/encryption.js';
 import { RECORDS_TABLE, RECORD_COUNTS_TABLE } from '../constants.js';
@@ -21,31 +20,10 @@ import { Err, Ok, retry, stringToHash } from '@nangohq/utils';
 import type { Result } from '@nangohq/utils';
 import type { Knex } from 'knex';
 import { Cursor } from '../cursor.js';
-import deepmerge from '@fastify/deepmerge';
+import { deepMergeRecordData } from '../helpers/merge.js';
 
 dayjs.extend(utc);
 
-const merge = deepmerge({
-    // define algorithm for merging arrays - merges each item in the array
-    // rather than concatenating
-    mergeArray: (options) => {
-        const deepmerge = options.deepmerge;
-        const clone = options.clone;
-        return function (target, source) {
-            let i = 0;
-            const il = Math.max(target.length, source.length);
-            const result = new Array(il);
-            for (i = 0; i < il; ++i) {
-                if (i < source.length) {
-                    result[i] = deepmerge(target[i], source[i]);
-                } else {
-                    result[i] = clone(target[i]);
-                }
-            }
-            return result;
-        };
-    }
-});
 const BATCH_SIZE = 1000;
 
 export async function getRecordCountsByModel({
@@ -473,7 +451,7 @@ export async function update({
 
                     const newRecord: FormattedRecord = {
                         ...newRecordRest,
-                        json: merge(oldRecordData, newRecordData) as UnencryptedRecordData,
+                        json: deepMergeRecordData(oldRecordData, newRecordData),
                         updated_at: new Date()
                     };
                     recordsToUpdate.push(newRecord);
